new functionality

+++ modules/channel.py
@@ -2,39 +2,46 @@ from telegram import Bot
 import random, os
 from .auth import check_token
 from .image_utils import image_is_correct
-from .exceptions import NotCorrectImage
+from .video_utils import video_is_correct
+from .exceptions import NotCorrectImage, NotCorrectVideo
 from loguru import logger
+
+
 class Channel:
     """
     The Channel class manages a single Telegram channel, allowing it to post images.
     """
 
 
-    def __init__(self, channel_id:str, bot_token:str, file_directory:str, file_caption:str=None) -> None:
+    def __init__(self, channel_id:str, bot_token:str, file_directory:str, file_caption:str=None, file_type:str='image', attempts = 20) -> None:
         self.channel_id = channel_id
         self.bot = self.__set_token(bot_token)
         self.file_directory = file_directory
         self.caption = self.__get_caption(file_caption)
+        self.file_type = file_type
+        self.attempts = attempts
 
-    async def __post(self, file_path:str) -> None:
-        """
-        Attempts to post an image to the channel. If the image is not correct, it raises a NotCorrectImage exception.
-        """
-
+    async def __post_file(self, file_type:str) -> None:
+        file_path = None
         try:
-            if image_is_correct(file_path):
-                with open(f'{file_path}', 'rb') as photo:
-                    await self.bot.send_photo(chat_id=f'{self.channel_id}', photo=photo, caption=self.caption)
-            else:
-                raise NotCorrectImage()
+            match file_type:
+                case 'image':
+                    file_path = self.__get_random_file('image')
+                    if image_is_correct(file_path):
+                        with open(f'{file_path}', 'rb') as photo:
+                            await self.bot.send_photo(chat_id=f'{self.channel_id}', photo=photo, caption=self.caption)
+                case 'video':
+                    file_path = self.__get_random_file('video')
+                    if video_is_correct(file_path):
+                        with open(f'{file_path}', 'rb') as video:
+                            await self.bot.send_video(chat_id=f'{self.channel_id}', video=video, caption=self.caption)
 
-        except Exception as e:
-            logger.error(f"An error occurred while trying to post the image: {e}")
-            raise
-
+        except Exception:
+            raise Exception
+        
         finally:
-            os.remove(file_path)
-
+            if file_path is not None:
+                os.remove(file_path)
 
     def __set_token(self, token:str) -> Bot:
         """
@@ -48,16 +55,30 @@ class Channel:
             raise ValueError("The bot token did not pass the check.")
 
 
-    def __get_random_file(self) -> str:
+    def __get_random_file(self, file_type) -> str:
+       
         """
         Returns a random file from the directory.
         If the directory is empty, raises an exception.
         """
+        img_extensions = ['jpeg', 'jpg', 'png', 'bmp', 'svg', 'webp']
+        video_extensions = ['mp4']
 
-        files = os.listdir(f'{self.file_directory}')
+        match file_type:
+            case 'image':
+                files = self.__get_files(img_extensions, self.file_directory)
+            case 'video':
+                files = self.__get_files(video_extensions, self.file_directory)
+
         if not files:
             raise Exception("The directory is empty.")
         return os.path.join(self.file_directory, random.choice(files))
+    
+    def __get_files(self, extensions, file_directory):
+        files = []
+        for extension in extensions:
+            files.extend(f for f in os.listdir(file_directory) if f.endswith('.' + extension))
+        return files
     
     def __get_caption(self, file_caption:str) -> str:
         """
@@ -67,14 +88,14 @@ class Channel:
         with open(f'{file_caption}', 'r') as file:
             return file.read()
 
-    async def run(self) -> None:
+    async def send_file(self) -> None:
         """
         Attempts to post an image to the channel up to 20 times. If all attempts fail, it raises an exception.
         """
 
-        for _ in range(20):
+        for _ in range(self.attempts):
             try:
-                await self.__post(self.__get_random_file())
+                await self.__post_file(self.file_type)  
                 break
 
             except NotCorrectImage:
